+++ src/pages/dashboard/Dashboard.tsx
@@ -1,11 +1,11 @@
 import React, { FC, ReactElement } from 'react';
 
-import Menu from 'components/menu/Menu';
+import { Menu } from 'components/menu/Menu';
 
 import './dashboard.scss';
 
 const Dashboard: FC = (): ReactElement => (
-  <div className='menu'>
+  <div className='main dashboard'>
     <div className='content-wrapper'>
       <Menu />
     </div>
